refactor: separates login page and create new user page
